backend/app/database.py
import os

from sqlalchemy import create_engine
from dotenv import load_dotenv
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker

load_dotenv()

DATABASE_URL = f"postgresql://{os.getenv('DB_USER', 'todouser')}:{os.getenv('DB_PASSWORD', 'todopass')}@{os.getenv('DB_HOST', 'localhost')}/{os.getenv('DB_NAME', 'tododb')}"

engine = create_engine(DATABASE_URL)

SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)

Base = declarative_base()

def get_db():
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()